# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCII Initial Scaffolding'"}],"paths":{"/loteria/apuesta":{"post":{"tags":["loteria-controller"],"operationId":"postApuesta","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApuestaDtoPost"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseApuestaDtoPost"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/loteria/total/{id_sorteo}":{"get":{"tags":["loteria-controller"],"operationId":"getInfoSorteoById","parameters":[{"name":"id_sorteo","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InfoApuestasEntity"}}}}}}},"/loteria/sorteo/{id_sorteo}":{"get":{"tags":["loteria-controller"],"operationId":"getSorteoById","parameters":[{"name":"id_sorteo","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Sorteo"}}}}}}},"/loteria/sorteo/":{"get":{"tags":["loteria-controller"],"operationId":"getSorteos","parameters":[{"name":"fecha","in":"query","required":false,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SorteoResponse"}}}}}}}}},"components":{"schemas":{"ApuestaDtoPost":{"type":"object","properties":{"fechaSorteo":{"type":"string","format":"date"},"idCliente":{"type":"string"},"numero":{"type":"string"},"montoApostado":{"type":"number","format":"double"}}},"ResponseApuestaDtoPost":{"type":"object","properties":{"idSorteo":{"type":"integer","format":"int32"},"fechaSorteo":{"type":"string","format":"date"},"idCliente":{"type":"string"},"numero":{"type":"string"},"resultado":{"type":"string"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"InfoApuestasEntity":{"type":"object","properties":{"idSorteo":{"type":"integer","format":"int32"},"fechaSorteo":{"type":"string"},"totalDeApuestas":{"type":"number","format":"double"},"totalPagado":{"type":"number","format":"double"},"totalEnReserva":{"type":"number","format":"double"}}},"ApuestasDelSorteo":{"type":"object","properties":{"idCliente":{"type":"string"},"numero":{"type":"string"},"resultado":{"type":"string"},"montoApostado":{"type":"number","format":"double"},"premio":{"type":"number","format":"double"}}},"Sorteo":{"type":"object","properties":{"idSorteo":{"type":"integer","format":"int32"},"fechaSorteo":{"type":"string"},"apuestas":{"type":"array","items":{"$ref":"#/components/schemas/ApuestasDelSorteo"}},"totalEnReserva":{"type":"number","format":"double"}}},"SorteoResponse":{"type":"object","properties":{"numeroSorteo":{"type":"integer","format":"int32"},"fecha":{"type":"string"},"dineroTotalAcumulado":{"type":"number","format":"double"},"numerosSorteados":{"type":"array","items":{"type":"array","items":{"type":"integer","format":"int32"}}}}}}}}